Crate common_query

Modules

columnar_value

error

function 🔒

logical_plan

physical_plan

prelude

signature 🔒

Signature module contains foundational types that are used to represent signatures, types, and return types of functions. Copied and modified from datafusion.

Structs

Output

new Output struct with output data(previously Output) and output meta

OutputMeta

OutputMeta stores meta information produced/generated during the execution

Enums

AddColumnLocation

OutputData

Original Output struct carrying result data to response/client/user interface

Traits

DfPhysicalPlan

ExecutionPlan represent nodes in the DataFusion Physical Plan.

Type Aliases

OutputCost

OutputRows