common_query::logical_plan

Module accumulator

Accumulator module contains the trait definition for aggregation function’s accumulators.

Structs

DfAccumulatorAdaptor

A wrapper type for our Accumulator to DataFusion’s Accumulator, so to make our Accumulator able to be executed by DataFusion query engine.

Traits

Accumulator

An accumulator represents a stateful object that lives throughout the evaluation of multiple rows and generically accumulates values.

AggrFuncTypeStore

AggrFuncTypeStore stores the aggregate function’s input data’s types.

AggregateFunctionCreator

An AggregateFunctionCreator dynamically creates Accumulator.

Functions

make_accumulator_function

make_return_function

make_state_function

Type Aliases

AggregateFunctionCreatorRef