common_query::logical_plan

Module expr

Functions

build_filter_from_timestamp

Builds an Expr that filters timestamp column from given timestamp range. Returns None if time range is None or full time range.

build_same_type_ts_filter

Builds a filter for a timestamp column with the same type as the timestamp column. Returns None if time range is None or full time range.

timestamp_to_literal 🔒

Converts a [Timestamp] to datafusion literal value.