common_query::logical_plan::expr

Function build_same_type_ts_filter

pub fn build_same_type_ts_filter(
    ts_schema: &ColumnSchema,
    time_range: Option<TimestampRange>,
) -> Option<Expr>

Builds a filter for a timestamp column with the same type as the timestamp column. Returns None if time range is None or full time range.