common_query::logical_plan

Function breakup_insert_plan

pub fn breakup_insert_plan(
    plan: &LogicalPlan,
    default_catalog: &str,
    default_schema: &str,
) -> Option<(TableName, Arc<LogicalPlan>)>

Convert a insert into logical plan to an (table_name, logical_plan) where table_name is the name of the table to insert into. logical_plan is the plan to be executed.

if input logical plan is not insert into table_name <input>, return None

Returned TableName will use provided catalog and schema if not specified in the logical plan, if table scan in logical plan have full table name, will NOT override it.