common_query::logical_plan

Function rename_logical_plan_columns

pub fn rename_logical_plan_columns(
    enable_ident_normalization: bool,
    plan: LogicalPlan,
    renames: Vec<(&str, &str)>,
) -> Result<LogicalPlan>

Rename columns by applying a new projection. Returns an error if the column to be renamed does not exist. The renames parameter is a Vector with elements in the form of (old_name, new_name).