Module common_query::signature

Signature module contains foundational types that are used to represent signatures, types, and return types of functions. Copied and modified from datafusion.

Structs

Signature

The Signature of a function defines its supported input types as well as its volatility.

Enums

TypeSignature

A function’s type signature, which defines the function’s supported argument types.

Volatility

A function’s volatility, which defines the functions eligibility for certain optimizations

Functions

concrete_types_to_arrow_types 🔒