common_query
Module function

Functions

make_scalar_function

decorates a function to handle ScalarValues by converting them to arrays before calling the function and vice-versa after evaluation.

to_df_return_type

Type Aliases

AccumulatorCreatorFunction

Create Accumulator with the data type of input columns.

AccumulatorFunctionImpl

Accumulator creator that will be used by DataFusion

ReturnTypeFunction

A function’s return type

ScalarFunctionImplementation

Scalar function

StateTypeFunction

This signature corresponds to which types an aggregator serializes its state, given its return datatype.