common_query

Module logical_plan

Modules

expr ๐

Traits

SubstraitPlanDecoder

The datafusion [LogicalPlan] decoder.

Functions

add_insert_to_logical_plan

create a insert into table_name <input> logical plan

breakup_insert_plan

Convert a insert into logical plan to an (table_name, logical_plan) where table_name is the name of the table to insert into. logical_plan is the plan to be executed.

build_filter_from_timestamp

Builds an Expr that filters timestamp column from given timestamp range. Returns None if time range is None or full time range.

build_same_type_ts_filter

Builds a filter for a timestamp column with the same type as the timestamp column. Returns None if time range is None or full time range.

rename_logical_plan_columns

Rename columns by applying a new projection. Returns an error if the column to be renamed does not exist. The renames parameter is a Vector with elements in the form of (old_name, new_name).

Type Aliases

SubstraitPlanDecoderRef