Module common_query::logical_plan

Re-exports

pub use self::accumulator::Accumulator;

pub use self::accumulator::AggregateFunctionCreator;

pub use self::accumulator::AggregateFunctionCreatorRef;

Modules

accumulator

Accumulator module contains the trait definition for aggregation function’s accumulators.

expr 🔒

udaf 🔒

Udaf module contains functions and structs supporting user-defined aggregate functions.

udf 🔒

Udf module contains foundational types that are used to represent UDFs. It’s modified from datafusion.

Structs

AggregateFunction

Logical representation of a user-defined aggregate function (UDAF) A UDAF is different from a UDF in that it is stateful across batches.

ScalarUdf

Logical representation of a UDF.

Traits

SubstraitPlanDecoder

The datafusion [LogicalPlan] decoder.

Functions

build_filter_from_timestamp

Builds an Expr that filters timestamp column from given timestamp range. Returns None if time range is None or full time range.

create_aggregate_function

create_udf

Creates a new UDF with a specific signature and specific return type. This is a helper function to create a new UDF. The function create_udf returns a subset of all possible ScalarFunction:

Type Aliases

SubstraitPlanDecoderRef