common_query::logical_plan::accumulator

Trait AggregateFunctionCreator

pub trait AggregateFunctionCreator: AggrFuncTypeStore {
    // Required methods
    fn creator(&self) -> AccumulatorCreatorFunction;
    fn output_type(&self) -> Result<ConcreteDataType>;
    fn state_types(&self) -> Result<Vec<ConcreteDataType>>;
}

An AggregateFunctionCreator dynamically creates Accumulator.

An AggregateFunctionCreator often has a companion struct, that can store the input data types (impl AggrFuncTypeStore), and knows the output and states types of an Accumulator.

Required Methods

fn creator(&self) -> AccumulatorCreatorFunction

Create a function that can create a new accumulator with some input data type.

fn output_type(&self) -> Result<ConcreteDataType>

Get the Accumulator’s output data type.

fn state_types(&self) -> Result<Vec<ConcreteDataType>>

Get the Accumulator’s state data types.