Trait common_query::physical_plan::PhysicalPlan

pub trait PhysicalPlan: Debug + Send + Sync {
    // Required methods
    fn as_any(&self) -> &dyn Any;
    fn schema(&self) -> SchemaRef;
    fn properties(&self) -> &PlanProperties;
    fn children(&self) -> Vec<PhysicalPlanRef>;
    fn with_new_children(
        &self,
        children: Vec<PhysicalPlanRef>
    ) -> Result<PhysicalPlanRef>;
    fn execute(
        &self,
        partition: usize,
        context: Arc<TaskContext>
    ) -> Result<SendableRecordBatchStream>;

    // Provided method
    fn metrics(&self) -> Option<MetricsSet> { ... }
}

PhysicalPlan represent nodes in the Physical Plan.

Each PhysicalPlan is partition-aware and is responsible for creating the actual “async” [SendableRecordBatchStream]s of RecordBatch that incrementally compute the operator’s output from its input partition.

Required Methods

fn as_any(&self) -> &dyn Any

Returns the physical plan as Any so that it can be downcast to a specific implementation.

fn schema(&self) -> SchemaRef

Get the schema for this physical plan

fn properties(&self) -> &PlanProperties

Return properties of the output of the PhysicalPlan, such as output ordering(s), partitioning information etc.

fn children(&self) -> Vec<PhysicalPlanRef>

Get a list of child physical plans that provide the input for this plan. The returned list will be empty for leaf nodes, will contain a single value for unary nodes, or two values for binary nodes (such as joins).

fn with_new_children( &self, children: Vec<PhysicalPlanRef> ) -> Result<PhysicalPlanRef>

Returns a new plan where all children were replaced by new plans. The size of children must be equal to the size of PhysicalPlan::children().

fn execute( &self, partition: usize, context: Arc<TaskContext> ) -> Result<SendableRecordBatchStream>

Creates an RecordBatch stream.

Provided Methods

fn metrics(&self) -> Option<MetricsSet>