common_query

Module prelude

Re-exports

pub use crate::columnar_value::ColumnarValue;

pub use crate::logical_plan::AggregateFunction;

Structs

Signature

The Signature of a function defines its supported input types as well as its volatility.

Enums

ScalarValue

A dynamically typed, nullable single value.

TypeSignature

A function’s type signature, which defines the function’s supported argument types.

Volatility

A function’s volatility, which defines the functions eligibility for certain optimizations

Constants

GREPTIME_COUNT

Default counter column name for OTLP metrics.

GREPTIME_PHYSICAL_TABLE

Default physical table name

GREPTIME_TIMESTAMP

Default timestamp column name for Prometheus metrics.

GREPTIME_VALUE

Default value column name for Prometheus metrics.

Functions

to_df_return_type

Type Aliases

AccumulatorCreatorFunction

Create Accumulator with the data type of input columns.

AccumulatorFunctionImpl

Accumulator creator that will be used by DataFusion

ReturnTypeFunction

A function’s return type

StateTypeFunction

This signature corresponds to which types an aggregator serializes its state, given its return datatype.